setup win python reqs
